refactor: lib/common/reset-repositories.js
