+++ lib/handle-installation.js
@@ -3,13 +3,10 @@ module.exports = handleInstallation
 const pluralize = require('pluralize')
 
 const getConfig = require('./app-config')
-const getPlan = require('./get-plan')
-const handlePullRequestChange = require('./handle-pull-request-change')
+const resetRepositories = require('./common/reset-repositories')
 
 // On install or accepted permissions, update all PRs for installs
 // On uninstall, just log
-//
-//
 async function handleInstallation (app, context) {
   const { action, repositories, repositories_added: added, repositories_removed: removed, installation: { account, repository_selection: selection } } = context.payload
 
@@ -27,83 +24,42 @@ async function handleInstallation (app, context) {
     return
   }
 
-  if (action === 'created') {
-    log.info(`🤗 ${account.type} ${account.login} installed on ${pluralize('repository', repositories.length, true)}`)
-    const repositoryNames = repositories.map(repository => repository.name)
-    const plan = await getPlan(app, account)
-    await Promise.all(repositoryNames.map(reset.bind(null, {
-      plan,
-      context,
-      owner: account.login
-    })))
+  if (action === 'removed') {
+    log.info(`➖ ${account.type} ${account.login} removed ${pluralize('repository', removed.length, true)}`)
     return
   }
 
-  if (action === 'new_permissions_accepted') {
-    log.info(`👌 ${account.type} ${account.login} accepted new permissions`)
+  if (action === 'created') {
+    log.info(`🤗 ${account.type} ${account.login} installed on ${pluralize('repository', repositories.length, true)}`)
 
-    const options = context.github.apps.listRepos.endpoint.merge({
-      per_page: 100
-    })
-
-    const repositories = await context.github.paginate(options)
-    const repositoryNames = repositories.map(repository => repository.name)
-    const plan = await getPlan(app, account)
-    await Promise.all(repositoryNames.map(reset.bind(null, {
-      plan,
+    return resetRepositories({
+      app,
       context,
-      owner: account.login
-    })))
-    return
+      account,
+      repositories
+    })
   }
 
   if (action === 'added') {
     log.info(`➕ ${account.type} ${account.login} added ${pluralize('repository', added.length, true)}`)
 
-    const repositoryNames = added.map(repository => repository.name)
-    const plan = await getPlan(app, account)
-
-    await Promise.all(repositoryNames.map(reset.bind(null, {
-      plan,
+    return resetRepositories({
+      app,
       context,
-      owner: account.login
-    })))
-    return
+      account,
+      repositories: added
+    })
   }
 
-  log.info(`➖ ${account.type} ${account.login} removed ${pluralize('repository', removed.length, true)}`)
-}
+  // action === new_permissions_accepted
+  log.info(`👌 ${account.type} ${account.login} accepted new permissions`)
 
-async function reset ({ plan, owner, context }, repo) {
-  const options = context.github.pulls.list.endpoint.merge({
-    owner,
-    repo,
-    state: 'open',
-    sort: 'updated',
-    direction: 'desc',
-    per_page: 100
+  const options = context.github.apps.listRepos.endpoint.merge({ per_page: 100 })
+
+  return resetRepositories({
+    app,
+    context,
+    account,
+    repositories: await context.github.paginate(options)
   })
-
-  const pullRequests = await context.github.paginate(options)
-  await Promise.all(
-    pullRequests.map(pullRequest => {
-      return handlePullRequestChange(null, {
-        plan,
-        github: context.github,
-        log: context.log,
-        repo (options) {
-          return Object.assign({
-            owner: owner,
-            repo
-          }, options)
-        },
-        event: 'pull_request',
-        payload: {
-          action: 'opened',
-          pull_request: pullRequest,
-          repository: pullRequest.base.repo
-        }
-      })
-    })
-  )
 }
